+++ src/main/java/simpleMoneyTransfer/webServices/api/accountWS/DeleteAccountWebService.java
@@ -41,7 +41,7 @@ public class DeleteAccountWebService {
     })
     public Response deleteAccount(
             @ApiParam(name = "Account Number", value = "Account Number for account to be deleted", required = true)
-            @QueryParam("accountNumber") Integer accountNumber,
+            @QueryParam("accountNumber") Long accountNumber,
             @ApiParam(name = "Accept-Language", value = "The value to be passed as header parameter",
                     required = true, defaultValue = "en-US")
             @HeaderParam("Accept-Language") @ValidLanguageCode String languageCode) {
