+++ src/main/java/simpleMoneyTransfer/webServices/api/accountWS/UpdateAccountWebService.java
@@ -48,7 +48,7 @@ public class UpdateAccountWebService {
     })
     public Response updateAccount(
             @ApiParam(name = "Account Number", value = "Account Number", required = true)
-            @QueryParam("accountNumber") Integer accountNumber,
+            @QueryParam("accountNumber") Long accountNumber,
             @ApiParam(name = "Accept-Language", value = "The value to be passed as header parameter",
                     required = true, defaultValue = "en-US")
             @HeaderParam("Accept-Language") @ValidLanguageCode String languageCode,
